refactor(joint-core): move removeEmptyAttributes to util module
Extract the recursive empty-attribute pruner and its top-level-only
predicate from `dia/Cell.mjs` into a self-contained
`util/removeEmptyAttributes.mjs`. Mirrors the `util/cloneCells.mjs`
pattern: imported directly by its single consumer, kept internal
(no entry in `util/index.mjs`).

Renames `isTopLevelEmpty` to `removeAtTopLevelOnly` so the constant's
pairing with `removeEmptyAttributes` is clear at the call site.

No behavior change.
